+++ api/Cargo.toml
@@ -28,10 +28,8 @@ reqwest = "0.10.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 uuid = "^0.7.0"
-# validator = "0.10.0"
-# validator_derive = "0.10.0"
-validator = { git = "https://github.com/LucasPickering/validator", branch = "blanket-ref-impl" }
-validator_derive = { git = "https://github.com/LucasPickering/validator", branch = "blanket-ref-impl" }
+validator = "0.10.1"
+validator_derive = "0.10.1"
 
 [dev-dependencies]
 maplit = "^1.0.0"
