+++ api/src/server/gql/hardware_spec.rs
@@ -9,13 +9,12 @@ use crate::{
         HardwareSpecConnectionFields, HardwareSpecEdgeFields,
         HardwareSpecNodeFields, PageInfo, UpdateHardwareSpecPayloadFields,
     },
-    util,
+    util::{self, Valid},
 };
 use diesel::{
     dsl, ExpressionMethods, OptionalExtension, PgConnection, QueryDsl,
     QueryResult, RunQueryDsl,
 };
-use gdlk::Valid;
 use juniper::ID;
 use juniper_from_schema::{QueryTrail, Walked};
 use std::convert::TryInto;
