+++ core/src/util.rs
@@ -4,9 +4,7 @@ use serde::{Deserialize, Serialize};
 use std::{
     fmt::{self, Formatter},
     iter,
-    ops::Deref,
 };
-use validator::{Validate, ValidationErrors};
 #[cfg(feature = "wasm")]
 use wasm_bindgen::prelude::*;
 
@@ -101,45 +99,6 @@ impl Span {
     /// a sub-slice of the given string that corresponds to this span.
     pub fn get_source_slice<'a>(&self, src: &'a str) -> &'a str {
         &src[self.offset..(self.offset + self.length)]
-    }
-}
-
-/// A small wrapper struct to indicate that the wrapped value has been
-/// validated. Built on top of [validator]. This struct can only be constructed
-/// via [Valid::validate].
-///
-/// ```
-/// use gdlk::{HardwareSpec, Valid};
-///
-/// let maybe_valid = HardwareSpec {
-///     num_registers: 1,
-///     num_stacks: 1,
-///     max_stack_length: 10,
-/// };
-/// let valid: Valid<HardwareSpec> = Valid::validate(maybe_valid).unwrap();
-/// ```
-#[derive(Copy, Clone, Debug)]
-pub struct Valid<T: Validate> {
-    inner: T,
-}
-
-impl<T: Validate> Valid<T> {
-    /// Validate the given value. If validation succeeds, wrap it in a
-    /// [Valid] to indicate it's valid.
-    pub fn validate(value: T) -> Result<Self, ValidationErrors> {
-        // We can't do a blanket TryFrom<T: Validate> implementation because of
-        // this bug https://github.com/rust-lang/rust/issues/50133
-        // Will have to wait for better specialization
-        value.validate()?;
-        Ok(Self { inner: value })
-    }
-}
-
-impl<T: Validate> Deref for Valid<T> {
-    type Target = T;
-
-    fn deref(&self) -> &T {
-        &self.inner
     }
 }
